+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [these versioning and changelog guidelines][changelog-gui
 
 ### Enhancements
 
+- Added support for complex operators in `ResourceList` component [#2006](https://github.com/Shopify/polaris-react/pull/2006)
 - Updated the page component's primary action to support `Button` props. ([#1994](https://github.com/Shopify/polaris-react/pull/1994))
 
 ## 2.6.1 - 2018-08-21